Rename EXT_mesh_primitive_restart -> KHR_mesh_primitive_restart
