+++ packages/extensions/src/constants.ts
@@ -1,5 +1,4 @@
 export const EXT_MESH_GPU_INSTANCING = 'EXT_mesh_gpu_instancing';
-export const EXT_MESH_PRIMITIVE_RESTART = 'EXT_mesh_primitive_restart';
 export const EXT_MESHOPT_COMPRESSION = 'EXT_meshopt_compression';
 export const EXT_TEXTURE_WEBP = 'EXT_texture_webp';
 export const EXT_TEXTURE_AVIF = 'EXT_texture_avif';
@@ -21,6 +20,7 @@ export const KHR_MATERIALS_TRANSMISSION = 'KHR_materials_transmission';
 export const KHR_MATERIALS_UNLIT = 'KHR_materials_unlit';
 export const KHR_MATERIALS_VOLUME = 'KHR_materials_volume';
 export const KHR_MATERIALS_VARIANTS = 'KHR_materials_variants';
+export const KHR_MESH_PRIMITIVE_RESTART = 'KHR_mesh_primitive_restart';
 export const KHR_MESH_QUANTIZATION = 'KHR_mesh_quantization';
 export const KHR_NODE_VISIBILITY = 'KHR_node_visibility';
 export const KHR_TEXTURE_BASISU = 'KHR_texture_basisu';
