+++ packages/extensions/src/index.ts
@@ -1,6 +1,5 @@
 import type { Extension } from '@gltf-transform/core';
 import { EXTMeshGPUInstancing } from './ext-mesh-gpu-instancing/index.js';
-import { EXTMeshPrimitiveRestart } from './ext-mesh-primitive-restart/index.js';
 import { EXTMeshoptCompression } from './ext-meshopt-compression/index.js';
 import { EXTTextureAVIF } from './ext-texture-avif/index.js';
 import { EXTTextureWebP } from './ext-texture-webp/index.js';
@@ -22,6 +21,7 @@ import { KHRMaterialsTransmission } from './khr-materials-transmission/index.js'
 import { KHRMaterialsUnlit } from './khr-materials-unlit/index.js';
 import { KHRMaterialsVariants } from './khr-materials-variants/index.js';
 import { KHRMaterialsVolume } from './khr-materials-volume/index.js';
+import { KHRMeshPrimitiveRestart } from './khr-mesh-primitive-restart/index.js';
 import { KHRMeshQuantization } from './khr-mesh-quantization/index.js';
 import { KHRNodeVisibility } from './khr-node-visibility/node-visibility.js';
 import { KHRTextureBasisu } from './khr-texture-basisu/index.js';
@@ -47,6 +47,7 @@ export const KHRONOS_EXTENSIONS: (typeof Extension)[] = [
 	KHRMaterialsUnlit,
 	KHRMaterialsVariants,
 	KHRMaterialsVolume,
+	KHRMeshPrimitiveRestart,
 	KHRMeshQuantization,
 	KHRNodeVisibility,
 	KHRTextureBasisu,
@@ -56,7 +57,6 @@ export const KHRONOS_EXTENSIONS: (typeof Extension)[] = [
 
 export const ALL_EXTENSIONS: (typeof Extension)[] = [
 	EXTMeshGPUInstancing,
-	EXTMeshPrimitiveRestart,
 	EXTMeshoptCompression,
 	EXTTextureAVIF,
 	EXTTextureWebP,
@@ -64,7 +64,6 @@ export const ALL_EXTENSIONS: (typeof Extension)[] = [
 ];
 
 export * from './ext-mesh-gpu-instancing/index.js';
-export * from './ext-mesh-primitive-restart/index.js';
 export * from './ext-meshopt-compression/index.js';
 export * from './ext-texture-avif/index.js';
 export * from './ext-texture-webp/index.js';
@@ -86,6 +85,7 @@ export * from './khr-materials-transmission/index.js';
 export * from './khr-materials-unlit/index.js';
 export * from './khr-materials-variants/index.js';
 export * from './khr-materials-volume/index.js';
+export * from './khr-mesh-primitive-restart/index.js';
 export * from './khr-mesh-quantization/index.js';
 export * from './khr-node-visibility/index.js';
 export * from './khr-texture-basisu/index.js';
