+++ packages/extensions/src/khr-mesh-primitive-restart/index.ts
