+++ packages/extensions/test/mesh-primitive-restart.test.ts
@@ -1,5 +1,5 @@
 import { Document, type JSONDocument, NodeIO } from '@gltf-transform/core';
-import { EXTMeshPrimitiveRestart } from '@gltf-transform/extensions';
+import { KHRMeshPrimitiveRestart } from '@gltf-transform/extensions';
 import { cloneDocument } from '@gltf-transform/functions';
 import test from 'ava';
 
@@ -7,12 +7,12 @@ const WRITER_OPTIONS = { basename: 'extensionTest' };
 
 test('basic', async (t) => {
 	const document = new Document();
-	const primRestartExtension = document.createExtension(EXTMeshPrimitiveRestart);
+	const primRestartExtension = document.createExtension(KHRMeshPrimitiveRestart);
 
 	let jsonDoc: JSONDocument;
 
-	jsonDoc = await new NodeIO().registerExtensions([EXTMeshPrimitiveRestart]).writeJSON(document, WRITER_OPTIONS);
-	t.deepEqual(jsonDoc.json.extensionsUsed, [EXTMeshPrimitiveRestart.EXTENSION_NAME], 'writes extensionsUsed');
+	jsonDoc = await new NodeIO().registerExtensions([KHRMeshPrimitiveRestart]).writeJSON(document, WRITER_OPTIONS);
+	t.deepEqual(jsonDoc.json.extensionsUsed, [KHRMeshPrimitiveRestart.EXTENSION_NAME], 'writes extensionsUsed');
 
 	primRestartExtension.dispose();
 
@@ -22,7 +22,7 @@ test('basic', async (t) => {
 
 test('copy', (t) => {
 	const doc = new Document();
-	doc.createExtension(EXTMeshPrimitiveRestart);
+	doc.createExtension(KHRMeshPrimitiveRestart);
 
-	t.is(cloneDocument(doc).getRoot().listExtensionsUsed().length, 1, 'copy EXTMeshPrimitiveRestart');
+	t.is(cloneDocument(doc).getRoot().listExtensionsUsed().length, 1, 'copy KHRMeshPrimitiveRestart');
 });
